apps/fatesclaw-dashboard/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fatesclaw-dashboard"
version = "0.1.0"
description = "OLED dashboard and control surface for OpenClaw on Raspberry Pi 4B with Fates hardware"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "fatesClaw contributors" }]
dependencies = [
  "websockets>=12,<16",
  "luma.oled>=3.13.0",
  "Pillow>=10.0.0",
  "gpiozero>=2.0.1",
  "evdev>=1.7.1",
]

[project.optional-dependencies]
dev = []

[project.scripts]
fatesclaw-dashboard = "fatesclaw_dashboard.main:run"

[tool.setuptools]
package-dir = { "" = "src" }
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["fatesclaw_dashboard*"]

[tool.ruff]
line-length = 100
target-version = "py311"
